Write to the assets file if there are any differences
In development I don't use any hash appended to the file, so the JSON file with the asset that is generated is very similar and doesn't change much.
Using this approach, I only write to the file if needed.

Since I use symfony to read the assets from the YML file. If the file has a different modification date, he tries to load again the assets, even if they are the same.