Add link to webpack-subresource-integrity package
I went down a bit of a rabbit hole trying to figure out why this wasn't working for me today, it turns out you need to use another plugin to make it work. By making this text a link, it's a little more clear that it depends on another plugin.